don't change number of replicas when upgrading
