+++ bin/hkube
@@ -5,7 +5,7 @@ const pull = require('../pull');
 const npm = require('../npm');
 const minikube = require('../minikube/minikube');
 const pack = require('../deployment/pack')
-const deploy = require('../deployment/deploy')
+const {deploy} = require('../deployment/deploy')
 const loadHkube = require('../deployment/load');
 const deleteHkube = require('../deployment/delete')
 const package = require('../package.json')
