+++ package.json
@@ -24,7 +24,7 @@
     "github-repositories": "^3.0.0",
     "js-yaml": "^3.10.0",
     "json-stream": "^1.0.0",
-    "kubernetes-client": "^3.16.1",
+    "kubernetes-client": "^5.2.0",
     "lodash": "^4.17.5",
     "meow": "^3.7.0",
     "minimist": "^1.2.0",
